--- src/Serenity.AIHub.SDK.NET.Core/Models/VolatileKnowledge/VolatileKnowledgeSimpleStatus.cs
@@ -1,10 +1,0 @@
-namespace Serenity.AIHub.SDK.NET.Core.Models.VolatileKnowledge;
-
-public static class VolatileKnowledgeSimpleStatus
-{
-    public const string Analyzing = "analyzing";
-    public const string Invalid = "invalid";
-    public const string Success = "success";
-    public const string Error = "error";
-    public const string Expired = "expired";
-}
