+++ src/SerenityStar/Models/VolatileKnowledge/VolatileKnowledgeSimpleStatus.cs
@@ -1,0 +1,15 @@
+namespace SerenityStar.Models.VolatileKnowledge
+{
+    /// <summary>
+    /// Defines simple status values for volatile knowledge processing.
+    /// </summary>
+
+    public static class VolatileKnowledgeSimpleStatus
+    {
+        public const string Analyzing = "analyzing";
+        public const string Invalid = "invalid";
+        public const string Success = "success";
+        public const string Error = "error";
+        public const string Expired = "expired";
+    }
+}
